# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "CommentResponse": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "user": {
                        "type": "string"
                    },
                    "post": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "likes": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "replies": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "replyTo": {
                        "type": "string"
                    },
                    "createdAt": {
                        "oneOf": [
                            {
                                "format": "date",
                                "type": "string"
                            },
                            {
                                "format": "date-time",
                                "type": "string"
                            }
                        ]
                    },
                    "updatedAt": {
                        "oneOf": [
                            {
                                "format": "date",
                                "type": "string"
                            },
                            {
                                "format": "date-time",
                                "type": "string"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "user",
                    "post",
                    "content",
                    "likes",
                    "replies",
                    "replyTo",
                    "createdAt",
                    "updatedAt"
                ]
            },
...
            "CommentOnPostRequest": {
                "properties": {
                    "content": {
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 5
                    }
                },
                "type": "object",
                "required": [
                    "content"
                ]
            }
        }
    },
    "info": {
        "title": "SocialApp API",
        "version": "1.0.0",
        "description": "Generated with `routing-controllers-openapi`"
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/posts/{postId}/comment": {
            "post": {
                "operationId": "PostController.commentOnPost",
                "parameters": [
                    {
                        "in": "path",
                        "name": "postId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentOnPostRequest"
                            }
                        }
                    },
                    "description": "CommentOnPostRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Comment on a post",
                "tags": [
                    "Post"
                ],
                "description": "\n    Submit a comment on a post.\n    Minimum length of content is 5 characters and maximum is 1000 characters.\n    ",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/comments/{commentId}": {
            "patch": {
                "operationId": "CommentController.updateComment",
                "parameters": [
                    {
                        "in": "path",
                        "name": "commentId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentOnPostRequest"
                            }
                        }
                    },
                    "description": "CommentOnPostRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Update a comment",
                "tags": [
                    "Comment"
                ],
                "description": "\n    Update the content of a comment.\n    Minimum length of content is 5 characters and maximum is 1000 characters.\n    Updating the content is allowed within 30 minutes of submission.\n    ",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "operationId": "CommentController.deleteComment",
                "parameters": [
                    {
                        "in": "path",
                        "name": "commentId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Delete a comment",
                "tags": [
                    "Comment"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/comments/{commentId}/like": {
            "post": {
                "operationId": "CommentController.likeComment",
                "parameters": [
                    {
                        "in": "path",
                        "name": "commentId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Like a comment",
                "tags": [
                    "Comment"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/comments/{commentId}/reply": {
            "post": {
                "operationId": "CommentController.replyToComment",
                "parameters": [
                    {
                        "in": "path",
                        "name": "commentId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentOnPostRequest"
                            }
                        }
                    },
                    "description": "CommentOnPostRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Reply to a comment",
                "tags": [
                    "Comment"
                ],
                "description": "\n    Submit a reply to a comment.\n    Minimum length of content is 5 characters and maximum is 1000 characters.\n    User can't reply to a comment that is a reply to another comment.\n    ",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/comments/": {
            "get": {
                "operationId": "CommentController.getPostComments",
                "parameters": [
                    {
                        "in": "query",
                        "name": "postId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "lastDocumentId",
                        "schema": {
                            "pattern": "^[0-9a-fA-F]{24}$",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/CommentResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Get comments of a post",
                "tags": [
                    "Comment"
                ],
                "description": "\n    Get comments of a post paginated and sorted by the creation date in ascending order.\n    limit: limit of documents to return, default is 5.\n    lastDocumentId: id of the last document in the previous page. If not provided, it will return the first page.\n    ",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/comments/{commentId}/replies": {
            "get": {
                "operationId": "CommentController.getCommentReplies",
                "parameters": [
                    {
                        "in": "path",
                        "name": "commentId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "lastDocumentId",
                        "schema": {
                            "pattern": "^[0-9a-fA-F]{24}$",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/CommentResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Get replies of a comment",
                "tags": [
                    "Comment"
                ],
                "description": "\n    Get replies of a comment paginated and sorted by the creation date in ascending order.\n    limit: limit of documents to return, default is 5.\n    lastDocumentId: id of the last document in the previous page. If not provided, it will return the first page.\n    ",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}